% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{province_stats_2024_63}
\alias{province_stats_2024_63}
\title{Provincial population and economic indicators under the 63-unit geography}
\format{
A data frame with 63 rows and 12 variables. Variables correspond to
those in [province_stats_2024], except that `iso` replaces `member_codes`.
}
\source{
General Statistics Office of Viet Nam, PX-Web tables, accessed
13 July 2026. \url{https://pxweb.nso.gov.vn/}
}
\usage{
province_stats_2024_63
}
\description{
The official 2024 source values before aggregation to Viet Nam's current
34-unit provincial geography.
}
\examples{
data(province_stats_2024_63)
plot_vnmap(
  geography = "provinces_63",
  data = province_stats_2024_63,
  values = "population_density",
  id = "code"
)
}
\keyword{datasets}
