% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vn-map.R
\name{vnmap_crs}
\alias{vnmap_crs}
\title{Coordinate reference system used by vnmap}
\usage{
vnmap_crs()
}
\value{
An \code{sf} CRS object.
}
\description{
EPSG:3405 is the VN-2000 / UTM zone 48N projected CRS.
}
\details{
All bundled geometries are returned in this projected CRS unless a
different value is supplied to the \code{crs} argument of \code{\link[=vn_map]{vn_map()}}. Using a
projected CRS keeps distances and polygon rendering stable for national
statistical maps.
}
\examples{
vnmap_crs()
}
\seealso{
\code{\link[=vn_map]{vn_map()}}, \code{\link[sf:st_crs]{sf::st_crs()}}
}
